Renaming ALogger to StatusLogger
