+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/libp2p/PeerManager.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.networking.p2p.libp2p;
 
-import static tech.pegasys.teku.logging.ALogger.STDOUT;
+import static tech.pegasys.teku.logging.StatusLogger.STDOUT;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Throwables;
