+++ services/beaconchain/src/main/java/tech/pegasys/artemis/services/beaconchain/BeaconChainService.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.services.beaconchain;
 
-import static tech.pegasys.teku.logging.ALogger.STDOUT;
+import static tech.pegasys.teku.logging.StatusLogger.STDOUT;
 
 import java.util.Objects;
 import org.apache.logging.log4j.Level;
