+++ sync/src/main/java/tech/pegasys/artemis/sync/AttestationManager.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.sync;
 
-import static tech.pegasys.teku.logging.ALogger.STDOUT;
+import static tech.pegasys.teku.logging.StatusLogger.STDOUT;
 
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
