+++ validator/coordinator/src/main/java/tech/pegasys/artemis/validator/coordinator/ValidatorLoader.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.validator.coordinator;
 
-import static tech.pegasys.teku.logging.ALogger.STDOUT;
+import static tech.pegasys.teku.logging.StatusLogger.STDOUT;
 
 import com.google.common.collect.Streams;
 import java.util.Collection;
